wifibt: Fix installation of wifibt scripts

+++ common/scripts/post-wifibt.sh
@@ -271,7 +271,9 @@ build_wifibt()
 	install -m 0655 $RKWIFIBT_DIR/conf/* "$TARGET_DIR/etc/"
 	install -m 0755 $RKWIFIBT_DIR/bin/arm/* "$TARGET_DIR/usr/bin/"
 	install -m 0755 $RKWIFIBT_DIR/scripts/* "$TARGET_DIR/usr/bin/"
-	rm -f "$TARGET_DIR/usr/bin/wifibt-sleep-hook.sh"
+	if [ -f "$RKWIFIBT_DIR/scripts/wifibt-sleep-hook.sh" ]; then
+		rm -f "$TARGET_DIR/usr/bin/wifibt-sleep-hook.sh"
+	fi
 	for b in bt-tty wifibt-info wifibt-vendor wifibt-id wifibt-bus \
 		wifibt-chip wifibt-module; do
 		ln -sf wifibt-util.sh "$TARGET_DIR/usr/bin/$b"
@@ -364,8 +366,10 @@ build_wifibt()
 	# Install suspend hook
 	for hook_dir in /usr/lib/pm-utils/sleep.d /lib/systemd/system-sleep; do
 		[ -d "$TARGET_DIR/$hook_dir" ] || continue
-		install -m 0755 $RKWIFIBT_DIR/scripts/wifibt-sleep-hook.sh \
-			"$TARGET_DIR/$hook_dir/03wifibt"
+		if [ -f "$RKWIFIBT_DIR/scripts/wifibt-sleep-hook.sh" ]; then
+			install -m 0755 $RKWIFIBT_DIR/scripts/wifibt-sleep-hook.sh \
+				"$TARGET_DIR/$hook_dir/03wifibt"
+		fi
 	done
 
 	# Log collection
